修改截图逻辑，支持保存截图的时候 0-10 循环下标保存最多11个截图文件（可自行修改）

+++ area_shot.sh
@@ -1,7 +1,18 @@
 #!/bin/bash
 
+echo "area_shot" > /home/busyo/文档/EnhanceScript/shot.txt
+
+FILE="/home/busyo/文档/EnhanceScript/shot_index.txt"
+
+if [ ! -f "$FILE" ]; then
+    echo "文件 $FILE 不存在，正在创建空文件..."
+    echo 0 > "$FILE"
+fi
+
+content=$(cat "$FILE")
+
 # 设置截图文件路径
-SCREENSHOT_PATH="/tmp/screenshot.png"
+SCREENSHOT_PATH="/tmp/screenshot${content}.png"
 
 # 删除旧的截图文件，确保 spectacle 生成新文件
 rm -f "$SCREENSHOT_PATH"
@@ -13,8 +24,6 @@ spectacle -b  -r  -o "$SCREENSHOT_PATH" &
 # 获取 spectacle 进程的PID
 SNIPASTE_PID=$!
 
-echo "area_shot" > /home/busyo/文档/EnhanceScript/shot.txt
-
 # 等待 spectacle 完成截图并退出
 wait $SNIPASTE_PID
 
@@ -23,9 +32,19 @@ sync
 
 echo "截图完成"
 
-# 复制截图到剪贴板
-echo -n "file://$SCREENSHOT_PATH" | wl-copy -t text/uri-list
+if [ -f "$SCREENSHOT_PATH" ]; then
+    echo "截图存在"
 
-echo "截图已复制到剪贴板 ✅"
+    # 复制截图到剪贴板
+    echo -n "file://$SCREENSHOT_PATH" | wl-copy -t text/uri-list
 
-echo "done" >  /home/busyo/文档/EnhanceScript/shot.txt
+    echo "截图已复制到剪贴板 ✅"
+
+    echo "done" >  /home/busyo/文档/EnhanceScript/shot.txt
+
+    ((content++))
+
+    if [ $content -gt 10 ]; then content=0; fi
+
+    echo $content > "$FILE"
+fi
